draw thread refactor
