+++ cli/util/mod.rs
@@ -5,6 +5,7 @@ pub mod checksum;
 pub mod console;
 pub mod diff;
 pub mod display;
+pub mod draw_thread;
 pub mod file_watcher;
 pub mod fs;
 pub mod logger;
